Updated installer.  Tests runs to completion.

+++ src/SpeciesData.cs
@@ -5,7 +5,7 @@ namespace Landis.Extension.OriginalFire
 {
     public class SpeciesData 
     {
-        public static Landis.Library.Parameters.Species.AuxParm<double> FireTolerance;
+        public static Landis.Library.Parameters.Species.AuxParm<int> FireTolerance;
 
         //---------------------------------------------------------------------
         public static void Initialize(IInputParameters parameters)
